Refactor: sc-27426 remove generate-assertions code from workspace.py
